Crate rinuxcore

Rinux

OS, written in rust

Basic Example:

#![no_std]
#![no_main]
#![feature(custom_test_frameworks)]
#![test_runner(rinuxcore::test_runner)]
#![reexport_test_harness_main = "test_main"]

use rinuxcore::{
    kernel, println,
    task::{
        executor::Executor,
        Task
    },
    BootInfo
};
kernel!(kernel_main);
fn kernel_main(boot_info: &'static BootInfo) -> ! {
    rinuxcore::init(boot_info);
    let mut executor = Executor::new();
 
    executor.spawn(Task::new(rinuxcore::task::keyboard::init()));
    executor.spawn(Task::new(main()));
 
    executor.run()
}
 
async fn main() {
    println!("Hello World");
}
 
#[cfg(not(test))]
#[panic_handler]
fn panic(info: &core::panic::PanicInfo) -> ! {
    rinuxcore::print_err!("{}", info);
    rinuxcore::hlt_loop();
}
#[cfg(test)]
#[panic_handler]
fn panic(info: &core::panic::PanicInfo) -> ! {
    rinuxcore::test_panic_handler(info)
}
#[test_case]
fn trivial_assertion() {
    assert_eq!(1, 1);
}

Re-exports

pub extern crate vga_buffer;

pub use x86_64;

pub use epearl;

Modules

allocator

conf

gdt

interrupts

memory

serial

task

Macros

kernel

print

print_err

println

serial_print

serial_println

Structs

BootInfo

This structure represents the information that the bootloader passes to the kernel.

Enums

ConfigType

Enum for Configuration Data Specification

QemuExitCode

Enum For Qemu Exit codes, sometimes useful

Traits

Testable

Useful for testing

Functions

exit_qemu

Quits Qemu using certain exit code

hlt_loop

Loop used to just do nothing

init

Initializes the core of Rinux

set_config_type

Used for setting source for rinux to get it’s config from

test_panic_handler

It’s in its name

test_runner

Runs Tests